OSINT/[4R0UND+H3GL083]/source/src/requirements.txt
Pillow
pygeohash
qrcode
